# perf(git): avoid unnecessary file hashing while config files detection