--- a/core/src/util/fs.ts
+++ b/core/src/util/fs.ts
@@ -135,6 +135,7 @@ export async function findConfigPathsInPath({
     log,
     filter: (f) => isConfigFilename(basename(f)),
     scanRoot: dir,
+    skipHashCalculation: true
   })
 
   return paths.map((f) => f.path)