--- a/core/src/util/fs.ts
+++ b/core/src/util/fs.ts
@@ -135,7 +135,7 @@ export async function findConfigPathsInPath({
     log,
     filter: (f) => isConfigFilename(basename(f)),
     scanRoot: dir,
-    skipHashCalculation: true,
+    hashUntrackedFiles: false,
   })
 
   return paths.map((f) => f.path)