--- a/core/src/vcs/vcs.ts
+++ b/core/src/vcs/vcs.ts
@@ -122,7 +122,7 @@ export interface GetFilesParams {
   filter?: (path: string) => boolean
   failOnPrompt?: boolean
   scanRoot: string | undefined
-  skipHashCalculation?: boolean
+  hashUntrackedFiles?: boolean
 }
 
 export interface BaseIncludeExcludeFiles {