**What this PR does / why we need it**:

As it was mentioned in #5920, Garden Core calls `AbstractGitHandler.getFiles()` method twice:

* While looking for Garden config files
* While actions/modules resolution

For big projects with a large amount of untracked files that can cause performance downgrade, because each calls trigger the hash calculation of the untracked files.

This PR makes a simple improvement by disabling hash-calculation in the config files detection phase.

It improves a performance only in a specific scenario, when too many files untracked by Git are included into the Garden.

**Which issue(s) this PR fixes**:

Fixes #

**Special notes for your reviewer**:

The performance difference can be compared by running `validate -l3`  Garden command in the root of https://github.com/garden-io/garden repo with modified https://github.com/garden-io/garden/blob/main/.gardenignore file. Delete `node_modules` from `.gardenignore` before running.

**Before**

```console
ℹ git                       → [verbose] Scanning project root at /Users/vladimirvagaytsev/Repositories/garden
  → Includes: (none)
  → Excludes: (none)
ℹ git                       → [verbose] Ignoring symlink with absolute target at /Users/vladimirvagaytsev/Repositories/garden/core/test/data/test-projects/build-dir/symlink-absolute/symlink.txt
ℹ git                       → [verbose] Found 127156 files in project root /Users/vladimirvagaytsev/Repositories/garden (took 9.98 sec)
...
ℹ graph                     → Resolving actions and modules...
ℹ git                       → [verbose] Scanning module e2e-tests root at /Users/vladimirvagaytsev/Repositories/garden
  → Includes: (none)
  → Excludes: (none)
ℹ git                       → [verbose] Found 127156 files in module e2e-tests root /Users/vladimirvagaytsev/Repositories/garden (took 9.97 sec)
⚠ graph                     → Large number of files (127156) found in module e2e-tests. You may need to configure file exclusions.
See https://docs.garden.io/using-garden/configuration-overview#including-excluding-files-and-directories for details.
✔ graph                     → Finished resolving graph (took 13.4 sec)
ℹ garden                    → Finished initializing Garden (took 25.1 sec)
```

**After**

```console
ℹ git                       → [verbose] Scanning project root at /Users/vladimirvagaytsev/Repositories/garden
  → Includes: (none)
  → Excludes: (none)
ℹ git                       → [verbose] Ignoring symlink with absolute target at /Users/vladimirvagaytsev/Repositories/garden/core/test/data/test-projects/build-dir/symlink-absolute/symlink.txt
ℹ git                       → [verbose] Found 127156 files in project root /Users/vladimirvagaytsev/Repositories/garden (took 1.73 sec)
...
ℹ graph                     → Resolving actions and modules...
ℹ git                       → [verbose] Scanning module e2e-tests root at /Users/vladimirvagaytsev/Repositories/garden
  → Includes: (none)
  → Excludes: (none)
ℹ git                       → [verbose] Found 127156 files in module e2e-tests root /Users/vladimirvagaytsev/Repositories/garden (took 10.79 sec)
⚠ graph                     → Large number of files (127156) found in module e2e-tests. You may need to configure file exclusions.
See https://docs.garden.io/using-garden/configuration-overview#including-excluding-files-and-directories for details.
✔ graph                     → Finished resolving graph (took 15.8 sec)
ℹ garden                    → Finished initializing Garden (took 19.5 sec)
```