@stefreak 2024-09-23
@vvagaytsev in case the large number of untracked files are included in an action, the user will get a warning 
```
Large number of files (127156) found in module e2e-tests. You may need to configure file exclusions.
See https://docs.garden.io/using-garden/configuration-overview#including-excluding-files-and-directories for details.
```

So how do we hide a problem? If the user doesn't want those files to be included, they need to configure the action not to include them.

Once we get to the point that we only calculate the hash if the file is included in the action, it doesn't matter how the user excludes the file (using the gardenignore feature or includes/excludes).

In the end, we can't know if the files need to be excluded or included and so we shouldn't send the user to a potentially wrong direction. If the files are not included, they shouldn't slow down Garden. If they are included, the user should indeed be informed that it might slow down things.