@vvagaytsev 2024-09-24
@stefreak yes, you're right.
We do have a warning printed on the large number of included files, both tracked and untracked. I don't think it makes any sense to introduce one more warning for untracked files only.

Regarding the problem "hiding", sorry for the confusion, let me elaborate that a bit.

I meant that the changes made in this PR can "hide" a bigger problem. This PR is not the proper solution, but just a quick-fix to reduce the processing time and make it not-too-slow for "some" (not too large) number of untracked files. Maybe "deferring a problem" would be a better wording here :)

This PR covers a specific case, but does not fix the performance issue for a "large enough" number of untracked files.
Thus, I'm not sure if it's a good idea to merge this.