@stefreak 2024-09-24
@vvagaytsev I think it might make sense to change the warning to only warn you if there is a large number of untracked files, if that's the only case that can cause a performance issue; Definitely open to think about that separately 👍 

It feels to me that this PR gives a tangible performance improvement already, and while it does not solve the problem entirely it does not make it worse or hide it so I am pro merging this 👍