@stefreak 2024-09-24
There is a test failure though, maybe avoiding the hashing in the first pass is causing a regression?