@vvagaytsev 2024-09-24
@stefreak right, actually, we should notify users on the large number of untracked and modified files, i.e. the files we need to calculate hashes for.
Having 10000+ is not a problem if all of those are committed to Git, because the hashes are already calculated.

I'll take a closer look at the test failures.