@vvagaytsev 2024-09-24
@stefreak the tests should be green now. Please review. I'll change the warning on the large number of untracked files in a separate PR.