@vvagaytsev 2024-09-24
> Are we now using the repo scan mode when scanning for configs? If we are, I think we may want to address that, since scanning for configs is otherwise quite efficient and shouldn't trigger any file hashing at all.

@edvald yes, we use `repo` scan mode. The config files scanner uses the same VCS handler instance as the Garden does, and the `repo` scan mode has been a default one since 0.13.20.