@edvald 2024-09-24
Got it. I think we should look into that, could be a performance drag when scanning for configs in many cases.