@edvald 2024-09-25
Agreed, let's do that in a separate PR. But I'm quite sure that using the other scan mode to find config files will work better, at the very least in commands that don't need to resolve all or many actions, where the repo is fairly large.